requirements.txt
Django>=2.1.11
gunicorn==19.9.0

djangorestframework
psycopg2
pygments
include
django-cors-headers